+++ src_hw/layers/pooling_layer.hpp
@@ -12,11 +12,11 @@ public:
 	PoolingLayer(std::string name, std::string bottom, std::string top, int pad, int kernelSize, int stride);
 	virtual ~PoolingLayer(){}
 
-	virtual void LayerSetUp(const Blob<float>* bottom, const Blob<float>* top);
+	virtual void LayerSetUp(const Blob<int>* bottom, const Blob<int>* top);
 
-	virtual void Reshape(const Blob<float>* bottom, Blob<float>* top);
+	virtual void Reshape(const Blob<int>* bottom, Blob<int>* top);
 
-	virtual void Forward(const Blob<float>* bottom, Blob<float>* top);
+	virtual void Forward(const Blob<int>* bottom, Blob<int>* top);
 
 	virtual inline const int type() const
 	{
