+++ src_hw/util/maths_functions.hpp
@@ -7,7 +7,7 @@ void conv_cpu(int stride, int pad,
 		int outColStart, int outColEnd,
 		int outDepthStart, int outDepthEnd,
 		int inDepthStart, int inDepthEnd,
-		const float* input, const float* weights, float* output
+		const int* input, const int* weights, int* output
 );
 
 //#pragma SDS data copy(inputTile[ 0 : IN_DEPTH_TILE_SIZE_3X3 * inRowTileSize * inColTileSize ], weigthsTile[ 0 : OUT_DEPTH_TILE_SIZE_3X3 * IN_DEPTH_TILE_SIZE_3X3 * KERNEL_SIZE_3X3 * KERNEL_SIZE_3X3 ])
@@ -18,9 +18,9 @@ void conv_cpu(int stride, int pad,
 //		int outColStart, int outColEnd,
 //		int outDepthStart, int outDepthEnd,
 //		int inDepthStart, int inDepthEnd,
-//		float inputTile [ IN_DEPTH_TILE_SIZE_3X3 * IN_ROW_TILE_SIZE_3X3 * IN_COL_TILE_SIZE_3X3 ],
-//		float weightsTile [ OUT_DEPTH_TILE_SIZE_3X3 * IN_DEPTH_TILE_SIZE_3X3 * KERNEL_SIZE_3X3 * KERNEL_SIZE_3X3 ],
-//		float outputTile [ OUT_DEPTH_TILE_SIZE_3X3 * OUT_ROW_TILE_SIZE_3X3 * OUT_COL_TILE_SIZE_3X3 ]
+//		int inputTile [ IN_DEPTH_TILE_SIZE_3X3 * IN_ROW_TILE_SIZE_3X3 * IN_COL_TILE_SIZE_3X3 ],
+//		int weightsTile [ OUT_DEPTH_TILE_SIZE_3X3 * IN_DEPTH_TILE_SIZE_3X3 * KERNEL_SIZE_3X3 * KERNEL_SIZE_3X3 ],
+//		int outputTile [ OUT_DEPTH_TILE_SIZE_3X3 * OUT_ROW_TILE_SIZE_3X3 * OUT_COL_TILE_SIZE_3X3 ]
 //);
 
 void conv_cpu_transB(int stride, int pad,
@@ -29,9 +29,9 @@ void conv_cpu_transB(int stride, int pad,
 		int outColStart, int outColEnd,
 		int outDepthStart, int outDepthEnd,
 		int inDepthStart, int inDepthEnd,
-		const float* input, const float* weights, float* output
+		const int* input, const int* weights, int* output
 );
 
-void gemm_cpu(const bool isTransposeA, const bool isTransposeB, const int m, const int n, const int k, const float alpha, const float* A, const float* B, const float beta, float* C);
+void gemm_cpu(const bool isTransposeA, const bool isTransposeB, const int m, const int n, const int k, const int alpha, const int* A, const int* B, const int beta, int* C);
 
 #endif /* SRC_UTIL_MATHS_FUNCTIONS_HPP_ */
