+++ src_sw/layers/dropout_layer.hpp
@@ -15,9 +15,9 @@ public:
 	DropoutLayer(std::string name, std::string bottom, std::string top);
 	virtual ~DropoutLayer(){}
 
-	virtual void Reshape(const Blob<float>* bottom, Blob<float>* top);
+	virtual void Reshape(const Blob<int>* bottom, Blob<int>* top);
 
-	virtual void Forward(const Blob<float>* bottom, Blob<float>* top);
+	virtual void Forward(const Blob<int>* bottom, Blob<int>* top);
 
 	virtual inline const int type() const
 	{
